refactor: moved company addresses to company package
