+++ src/main/java/com/pawelnu/projectmanager/config/DataInit.java
@@ -7,8 +7,8 @@ import com.pawelnu.projectmanager.endpoints.authority.EmployeeAuthorityEntity;
 import com.pawelnu.projectmanager.endpoints.authority.EmployeeAuthorityRepository;
 import com.pawelnu.projectmanager.endpoints.company.CompanyEntity;
 import com.pawelnu.projectmanager.endpoints.company.CompanyRepository;
-import com.pawelnu.projectmanager.endpoints.companyaddress.CompanyAddressEntity;
-import com.pawelnu.projectmanager.endpoints.companyaddress.CompanyAddressRepository;
+import com.pawelnu.projectmanager.endpoints.company.address.CompanyAddressEntity;
+import com.pawelnu.projectmanager.endpoints.company.address.CompanyAddressRepository;
 import com.pawelnu.projectmanager.endpoints.employee.EmployeeEntity;
 import com.pawelnu.projectmanager.endpoints.employee.EmployeeRepository;
 import com.pawelnu.projectmanager.enums.CompanyStatus;
