+++ src/main/java/com/pawelnu/projectmanager/endpoints/company/address/CompanyAddressDTO.java
@@ -1,4 +1,4 @@
-package com.pawelnu.projectmanager.endpoints.companyaddress;
+package com.pawelnu.projectmanager.endpoints.company.address;
 
 import java.util.UUID;
 import lombok.Builder;
