+++ src/main/java/com/pawelnu/projectmanager/endpoints/company/address/CompanyAddressEditRequestDTO.java
@@ -1,14 +1,12 @@
-package com.pawelnu.projectmanager.endpoints.companyaddress;
+package com.pawelnu.projectmanager.endpoints.company.address;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
-import java.util.UUID;
 import lombok.Data;
 
 @Data
-public class CompanyAddressCreateRequestDTO {
+public class CompanyAddressEditRequestDTO {
 
-  @NotNull private UUID companyId;
   @NotNull private String street;
   @NotNull private String streetNumber;
   @NotNull private String city;
