+++ src/main/java/com/pawelnu/projectmanager/endpoints/company/address/CompanyAddressRepository.java
@@ -1,4 +1,4 @@
-package com.pawelnu.projectmanager.endpoints.companyaddress;
+package com.pawelnu.projectmanager.endpoints.company.address;
 
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
